Josh and me did some more work. All the pieces for making a working game, save the collision checking if there.
TODO:
Collision
Random generation of obstacles
Polish creation of 3d objects (at the moment there's allot of repeated code)
Accelerometer controls and a tilt to the movement of the ship (tilt phone left, ship tilts a little to the left)
Energy bar
*NOT A COMPREHENSIVE LIST

The way the code works is that the ship and tunnel are stationary, while the obstacles move towards the screen. Movement of a GLGameObject is controlled in the constructor, where you can set the initial x, y and z speeds.